Interface and Implementation of TimeSeriesDao.
